requirements.txt
streamlit
st-supabase-connection
python-dotenv